Run our deploy workflow in all the situations where we run our check workflow
The previous logic didn't make any sense: this workflow would only run
against master, but then it would check if we're on a tag before
deploying to pypi! I don't think it would be possible for both of those
to be true simultaneously (I know it's possible for a single commit hash
to be both `master` and a tag at the same time, but I think those would
each get considered separately by github for a workflow run)

The new logic is: let's run this in the exact same situations where we
run our check workflow, even PRs. Obviously we wouldn't want to deploy
anything from a PR, but that's fine: we've got if statements that make
sure we only deploy to pypi from our chosen special branches. And no,
there's no risk of a malicious actor sending in a PR and editing
`.github/workflows/deploy.yml` to get us to release a version of their
malicious code to pypi.
https://github.com/jfly/shtuff/settings/secrets/actions says:

> Secrets are not passed to workflows that are triggered by a pull
> request from a fork. [Learn more](https://docs.github.com/actions/automating-your-workflow-with-github-actions/creating-and-using-encrypted-secrets).