Combine check and deploy into a single new workflow: ci
I realized that previously, we could deploy to pypi even if our tests
were failing! There was no dependency between our check and deploy
workflows. Some very light Googling leads me to believe that the right
way to express a dependency like this is between jobs in the same
workflow, rather than trying to do it across workflows.

Now we've got 4 jobs, and those can run a bit in parallel and depend on
each other. I *think* this will work.