test: fix flaky tests
For some reason, these tests starting failing on Python 3.11, I'm not
exactly sure why. These tests do seem to be making an incorrect
assumption that just because the text "exit" has been printed to the
screen that that means the process has actually exited. So, now we just
wait for the process to actually exit, and things seem to behave
reliably now.