Make development easier!
Nix shells are neat, thanks @jfly!

We stole a page from mycli on how to run this entry point / script
without having to pip install it.

Due to this, we just happened to notice a latent bug with this __version__
variable because we only get a version when it's installed.