Shtuff no longer requires shtuff to be in $PATH to work
Our nix build of shtuff was failing because it ran shtuff in a sandbox
that didn't have shtuff installed. We used to pip install it in the
Makefile but we opted out of doing that as it's unnecessary for Nix :)

Due to getting rid of the venv, we needed to invoke shtuff a bit
differently in the tests. That led to some interesting behavior.

As we no longer have access to `shtuff` (as it's not pip installed),
this meant the recursive `shtuff as` no longer works. Fixing this means
we no longer get two prompts (yay positive side effect!) which explains
the removal of the duplicate prompt assertion from some tests.

While we were debugging this, we made a few tests improvements. One of
these was moving from os.system() to subprocess.run() as os.system()
ignores the exit code, which made the test wait to timeout when looking
for output, as opposed to failing the test immediately.

We also moved from capture_output (which captures both stdout _and_
stderr) to only capturing stdout, which allows us to see the stderr when
debugging test failures.